fix: Added missing exception class and functions to public API (i.e. to `__all__`)
