+++ .github/workflows/constraints.txt
@@ -1,2 +1,2 @@
-tox==4.32.0
+tox==4.34.1
 tox-uv==1.29.0
