+++ .pre-commit-config.yaml
@@ -3,8 +3,6 @@ ci:
   autofix_prs: true
   autoupdate_schedule: quarterly
   autoupdate_commit_msg: 'chore(deps): pre-commit autoupdate'
-  skip:
-  - check-dependabot
 
 repos:
 - repo: https://github.com/pre-commit/pre-commit-hooks
@@ -28,14 +26,14 @@ repos:
   - id: check-readthedocs
 
 - repo: https://github.com/astral-sh/ruff-pre-commit
-  rev: v0.14.10
+  rev: v0.14.14
   hooks:
   - id: ruff-check
     args: [--fix, --exit-non-zero-on-fix, --show-fixes]
   - id: ruff-format
 
 - repo: https://github.com/astral-sh/uv-pre-commit
-  rev: "0.9.21"
+  rev: "0.9.26"
   hooks:
   - id: uv-lock
   - id: uv-sync
@@ -51,6 +49,6 @@ repos:
   - id: pyproject-fmt
 
 - repo: https://github.com/crate-ci/typos
-  rev: v1
+  rev: v1.42.1
   hooks:
   - id: typos
